refactor(ipc): moved the LoadClientSettingsIPCHandler
Moved the LoadClientSettingsIPCHandler to a deeper directory. The new directory structure allows us
to provider a better overview of all IPC handlers.
