+++ packages/client/src/ioc/container.ts
@@ -2,7 +2,7 @@ import { getConfiguredContainer } from '@botvy/framework/dist/ioc/container';
 import { IPCEventHandler } from '@botvy/framework/dist/ipc/IPCEventHandler';
 import { InitializationSide } from '@botvy/framework/dist/plugin/InitializationSide';
 
-import { LoadClientSettingsIPCHandler } from '../ipc/LoadClientSettingsIPCHandler';
+import { LoadClientSettingsIPCHandler } from '../ipc/settings/LoadClientSettingsIPCHandler';
 
 export const getClientContainer = async () => {
     const container = await getConfiguredContainer(InitializationSide.CLIENT);
